pyproject.toml
[tool.poetry]
name = "moodlenotif"
version = "0.1.0"
description = ""
authors = ["RuslanUC <dev_ruslan_uc@protonmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
tortoise-orm = "^0.20.0"
fastapi = "^0.104.1"
pyrogram = "^2.0.106"
tgcrypto = "^1.2.5"
aiohttp = "^3.9.1"
aiocron = "^1.8"
beautifulsoup4 = "^4.12.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

